refactoring

+++ SI/src/sigrun/context/Capability.cpp
@@ -20,7 +20,7 @@ Capability::~Capability()
 
 }
 
-const std::map<std::string, int> &Capability::capabilities() const
+const std::map<std::string, uint64_t> &Capability::capabilities() const
 {
     return d_capabilities;
 }
@@ -62,12 +62,12 @@ void Capability::remove_capability(const std::string &name)
     }
 }
 
-int Capability::num_capabilities() const
+uint32_t Capability::num_capabilities() const
 {
     return d_num_capabilities;
 }
 
-int Capability::consecutive_capability_id() const
+uint64_t Capability::consecutive_capability_id() const
 {
     return d_consecutive_capability_id;
 }
